Refactor check and common modules to utils.
