+++ tfrecorder/utils.py
@@ -14,8 +14,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Utilities for checking content of TFRecord files."""
+"""Miscellaneous utility functions."""
 
+from datetime import datetime
 from typing import Dict
 
 import csv
@@ -24,9 +25,8 @@ import os
 import tensorflow as tf
 
 from tfrecorder import beam_image
-from tfrecorder import common
-# TODO(cezequiel): Rename `dataset` module to something else
-from tfrecorder import dataset_loader as _dataset
+from tfrecorder import constants
+from tfrecorder import dataset_loader
 
 _OUT_IMAGE_TEMPLATE = 'image_{:0>3d}.png'
 
@@ -48,12 +48,12 @@ def _save_image_from_record(record: Dict[str, tf.Tensor], outfile: str):
   image.save(outfile)
 
 
-def check_tfrecords(
+def inspect(
     tfrecord_dir: str,
     split: str = 'TRAIN',
     num_records: int = 1,
     output_dir: str = 'output'):
-  """Checks TFRecords from a TFRecord directory generated by TFRecorder.
+  """Prints contents of TFRecord files generated by TFRecorder.
 
   Args:
     tfrecord_dir: TFRecord directory.
@@ -65,12 +65,12 @@ def check_tfrecords(
     `ValueError` when data for a given `split` could not be loaded.
   """
 
-  dataset = _dataset.load(tfrecord_dir).get(split)
+  dataset = dataset_loader.load(tfrecord_dir).get(split)
   if not dataset:
     raise ValueError(f'Could not load data for {split}')
 
   data_dir = os.path.join(
-      output_dir, 'check-tfrecords-' + common.get_timestamp())
+      output_dir, 'check-tfrecords-' + get_timestamp())
   os.makedirs(data_dir)
 
   with open(os.path.join(data_dir, 'data.csv'), 'wt') as f:
@@ -99,3 +99,21 @@ def check_tfrecords(
     print('Output written to {}'.format(data_dir))
 
     return data_dir
+
+
+def get_timestamp() -> str:
+  """Returns current date and time as formatted string."""
+  return datetime.now().strftime('%Y%m%d-%H%M%S')
+
+
+def copy_logfile_to_gcs(logfile: str, output_dir: str):
+  """Copies a logfile from local to gcs storage."""
+  try:
+    with open(logfile, 'r') as log_reader:
+      out_log = os.path.join(output_dir, constants.LOGFILE)
+      with tf.io.gfile.GFile(out_log, 'w') as gcs_logfile:
+        log = log_reader.read()
+        gcs_logfile.write(log)
+  except FileNotFoundError as e:
+    raise FileNotFoundError("Unable to copy log file {} to gcs.".format(
+        e.filename)) from e
